refactor between _utils and utils
